refactoring schedules into 1 file
