+++ Procfile
@@ -1,5 +1,4 @@
 bids: python -m tasks.track_bids
 sales: python -m tasks.track_phunkbot_sales
 nftx: python -m tasks.track_nftx_purchases
-sniper: python -m tasks.sniping_alerts
-phunkstats: python -m tasks.tweet_stats
+schedule: python -m tasks.schedule_tasks
