+++ controllers/stats.py
@@ -4,18 +4,16 @@ import json
 import operator
 import os
 import sys
-import time
 from collections import defaultdict
-from datetime import timedelta
+from datetime import timedelta, datetime, timezone
 
 import arrow
 import pymongo
-import schedule
 
 import settings
 from lib.web3_helpers.common.node import get_ens_domain_for_address, get_curated_nfts_holdings
 from utils.covalent import get_nft_unique_owners
-from utils.database import get_latest_stats, get_holder, update_holder, save_holder, save_stats
+from utils.database import get_latest_stats, get_holder, update_holder, save_holder, save_stats, fetch_sales
 from utils.twitter import tweet
 
 # TODO: This needs a lot of love... It was made in one go and barely touched ever since. A lot of room for errors,
@@ -28,15 +26,16 @@ mongo_client = pymongo.MongoClient(settings.MONGO_CONN_STRING)
 holders_coll = mongo_client['phunks']['holders']
 
 DEEP_FETCH = os.environ.get("DEEP_FETCH", True)  # change this to False to make things quicker
+TOP_SALES_COUNT = 4
 
 # script can be executed as cli for testing or one-time runs
 my_parser = argparse.ArgumentParser()
-my_parser.add_argument('-r', '--run', action="store_true")
 my_parser.add_argument('-s', '--silent', action="store_true")
+my_parser.add_argument('-a', '--aggregated', action="store_true")
 
 args = my_parser.parse_args()
-should_run_now = args.run
 should_tweet = not args.silent
+run_aggregated = args.aggregated
 
 with open(os.path.join("data", "curated.json")) as f:
     curated_contracts = json.loads(f.read())
@@ -246,16 +245,63 @@ def fetch_phunks_stats():
         tweet(tweet_content)
 
 
-if should_run_now:
-    fetch_phunks_stats()
-else:
-    # UTC times
-    schedule.every().day.at("05:00").do(fetch_phunks_stats)
-    schedule.every().day.at("17:00").do(fetch_phunks_stats)
+def get_aggregated_stats():
+    now = datetime.now(timezone.utc)
 
-    while True:
-        try:
-            schedule.run_pending()
-            time.sleep(60)
-        except Exception as e:
-            continue
+    filters = {
+        "created_at": {
+            "$gte": now - timedelta(hours=24),
+            "$lt": now
+        }
+    }
+    sales = list(fetch_sales(filters=filters))
+    sum_eth = 0
+    sum_usd = 0
+    top_text = "\n".join(
+        [
+            f"{index}. #{sale.get('token_id')} for Ξ{sale.get('eth_amount'):.2f} (${sale.get('usd_amount'):,.2f})"
+            for index, sale in enumerate(sales[:TOP_SALES_COUNT], start=1)])
+
+    top_images = [f"https://phunks.s3.us-east-2.amazonaws.com/notpunks/notpunk{str(sale.get('token_id')).zfill(4)}.png"
+                  for sale in sales[:TOP_SALES_COUNT]]
+
+    for sale in sales:
+        sum_eth += sale.get("eth_amount")
+        sum_usd += sale.get("usd_amount")
+
+    filters = {
+        "created_at": {
+            "$gte": now - timedelta(hours=48),
+            "$lt": now - timedelta(hours=24),
+        }
+    }
+
+    previous_24h_sales = list(fetch_sales(filters=filters))
+    previous_sum_eth = 0
+    previous_sum_usd = 0
+    for sale in previous_24h_sales:
+        previous_sum_eth += sale.get("eth_amount")
+        previous_sum_usd += sale.get("usd_amount")
+
+    eth_diff = sum_eth - previous_sum_eth
+    usd_diff = sum_usd - previous_sum_usd
+
+    tweet_text = f"""Last 24h stats 📈
+
+# of sales: {len(sales)}
+ETH volume: Ξ{sum_eth:.2f} ({f"+Ξ{eth_diff:.2f}" if eth_diff > 0 else f"-Ξ{abs(eth_diff):.2f}"})
+USD volume: ${sum_usd:,.2f} ({f"+${usd_diff:,.2f}" if usd_diff > 0 else f"-${abs(usd_diff):,.2f}"})
+
+Top sales:
+{top_text} 
+"""
+    print(tweet_text)
+    if should_tweet:
+        tweet(tweet_text, image_urls=top_images, file_extension="png")
+
+
+if __name__ == '__main__':
+    if run_aggregated:
+        get_aggregated_stats()
+    else:
+        fetch_phunks_stats()
